refactor: split views and widgets into separate files (#382)
Moved all view rendering functions from mod.rs into dedicated files:
- views/dashboard.rs, plan.rs, history.rs, events.rs, nodes.rs
- views/settings.rs, templates.rs, clarification.rs, diff.rs

Moved all widget rendering functions from widgets/mod.rs into dedicated files:
- widgets/dag_tree.rs, node_detail.rs, progress_bar.rs, status_bar.rs
- widgets/event_log.rs, cmd_bar.rs, metrics_panel.rs, types.rs

mod.rs now only contains the core run loop, event handling, and delegation.
widgets/mod.rs re-exports sub-modules and shared types.
views/mod.rs re-exports all view modules.

Build: clean | Clippy: clean | Tests: 27/27 | Format: clean
